AGRI-MITRA/backend/requirements.txt
flask
flask-cors
gunicorn
torch
torchvision
scikit-learn
numpy
pillow
joblib